enable a torchless version of package
Summary:
Remove the requirement of torch in PackageExporter and PackageImporter. Now both of these modules work without torch, however, they can use torch for the necessary custom operations like in the case of torchscript and torch.fx. We do this by extending two endpoints similar to pickle call `persistent_id` and` __reduce_package__`

Currently we have `package_exporter(_importer)` hold a `PackageExporter(Importer)` hold a vesion of itself which supports the `persistent_id` / `persistent_load` to save `torch` tensors

This is effectively just a part of the repimplentation of D35474185

Differential Revision: D35634346

fbshipit-source-id: 0fe6911a800f57b1ac700f869ccde4f9c71fd485