Deprecate _gate_torchscript_serialization (#5)
* add TARGETS to multipy package (#4)

Summary:
Pull Request resolved: https://github.com/pytorch/MultiPy/pull/4

This diff adds TARGETS to Multipy.package. As well as a lot of linting (sorry). For review I recommend primarily looking at the TARGETS files.

Differential Revision: D35566916

fbshipit-source-id: 13534c533c4f0c82e4dd8dbb467e5e18b8aeb68d

* Deprecate _gate_torchscript_serialization

Summary:
At this point it's confirmed that `_gate_torchscript_serialization` works given our integration with literay and usecases which use torchscript modules. Let's deprecate it here.

When we migrate everyone to this repo, however, we'll need to delete all uses of `_gate_torchscript_serialization`. T116999894

Reviewed By: d4l3k

Differential Revision: D35605117

fbshipit-source-id: ad1c31bd27900bff97c6f143379f5dd7a8d93473

Co-authored-by: sahanp <>