Add Generic ZipFile to package (#6)
* add TARGETS to multipy package (#4)

Summary:
Pull Request resolved: https://github.com/pytorch/MultiPy/pull/4

This diff adds TARGETS to Multipy.package. As well as a lot of linting (sorry). For review I recommend primarily looking at the TARGETS files.

Differential Revision: D35566916

fbshipit-source-id: bde21478a64895717b1a4d5b1b9b1a02dae2361f

* Deprecate _gate_torchscript_serialization

Summary:
At this point it's confirmed that `_gate_torchscript_serialization` works given our integration with literay and usecases which use torchscript modules. Let's deprecate it here.

When we migrate everyone to this repo, however, we'll need to delete all uses of `_gate_torchscript_serialization`. T116999894

Differential Revision: D35605117

fbshipit-source-id: 5218ced02d6dc6a4b1f52ecce72cc8ae6940a0d3

* Add Generic ZipFile to package

Summary:
Adds a generic zipfile / directory reader which does not have a dependency on torch into multipy.package. This allows us to create a torch::deploy/ multipy workflow which does not have a dependency on torch.

This is effectively just a part of the repimplentation of D35474185

Reviewed By: d4l3k

Differential Revision: D35629045

fbshipit-source-id: 86158874b7e2bff127f8be4ace93cc9fa676633d

Co-authored-by: sahanp <>