Update base for Update on "fixed paths to be general"
The culmination of these prs allows for MultiPy runtime (name pending) to be run as part of a repo separate of torch in order to provide MultiPy as its own library. 

Unfortunately, we have to export `torch_python_obj_static` which requires `pytorch` to have it's own version of Cpython. Future work would be to have both objects point to the same Cpython.

Instructions on how to build can be found in MultiPy/runtime/ReadMe.md


[ghstack-poisoned]