Simiplify build to remove manually building third-party dependencies (#57)
Summary:
Pull Request resolved: https://github.com/pytorch/multipy/pull/57

Ends up when we did `add_subdirectory(...)` we didn't need to actually build the library. This pr removes the action of doing that. Furthermore, it updates the readme with regards of how to build `multipy::runtime` and updates the `.gitignore` to include `.git` and the updated paths.

Test Plan: Imported from OSS

Reviewed By: d4l3k

Differential Revision: D37098420

Pulled By: PaliC

fbshipit-source-id: d39d62bdbb7d313121e659354225a56a0b69fce3