Update on "Add dynamic ABI to Multipy"
This PR creates a way to change the ABI of MultiPy. One simply needs to make use of the `ABI_EQUALS_1` macro to do it. To set ABI to 1 we can run `export ABI_EQUALS_1=ON` or when building Multipy run `cmake -DABI_EQUALS_1=ON ..` instead of `cmake ..`. Otherwise, the ABI should default to 0. 

We do this in order to be compatible with the binary distribution of pytorch which defaults to using an ABI which is equal to 1.

Limitations:
Currently tests can only be done when ABI=1 as the testing infrustructure
used by `ASSERT_TRUE` requires ABI=1 in Github actions, we should fix this!
The error message for this can be [found here](https://github.com/pytorch/multipy/runs/6981969659?check_suite_focus=true) by searching `test_deploy` and looking for the `undefined` errors.

[ghstack-poisoned]