Allow torch._C to be recognized a module in torch.package in pytorch/multipy (#91)
Summary:
Pull Request resolved: https://github.com/pytorch/multipy/pull/91

This pr addresses https://github.com/pytorch/multipy/issues/82 and https://github.com/pytorch/multipy/issues/44.

A C extension module behaves a bit differently than a normal python package as it does not contain a `__path__` attribute. However, these modules still have information about their submodules. This PR also checks if a module is a C extension module  and checks if the module we are looking for is in it's children.

For example, if we are importing `torch._C._nn` we check if the parent `torch._C` is a C extension module if necessary, and then check if `torch._C._nn` is a proper child of `torch._C`.

The corresponding PR in torch.package is https://github.com/pytorch/pytorch/pull/80917

Reviewed By: d4l3k

Differential Revision: D37630606

fbshipit-source-id: ee562573ea4c0bbc43facc5ae9a0fc4ad8f29cd6