multipy: add plugin_registry and plugin_torch (#103)
Summary:
Pull Request resolved: https://github.com/pytorch/multipy/pull/103

For dynamic loading we need to be able to load the interpreter without python or libtorch_python initially. Thus we need a plugin registry that we can use to late register the libtorch_python extension.

Load order:

* libinterpreter
* libpython
* libtorch_python
* plugin_torch.cpp

This adds the plugin registry but doesn't change the loading semantics -- a follow up diff will implement the dynamic loading.

The Optional.hpp changes are required since we're now using the move semantics which weren't used before in a torch context.

Reviewed By: PaliC

Differential Revision: D38093644

fbshipit-source-id: 8b0517097ad1ef78f36420da8092a37789a03929