multipy: pybind test (#88)
Summary:
Pull Request resolved: https://github.com/pytorch/multipy/pull/88

This enables a previously broken (never working?) pybind test for multipy. This test and changes ensure that multipy/deploy can be pybinded into Python. The main gaps were that MultiPy was looking for the zipped sections/intepreter in the main executable where as now it searches all dynamically loaded libraries for the sections.

This might be slightly slower than when mulitpy is linked into the main binary since it has to mmap and search all loaded elf files but shouldn't be too bad as it only has to happen at initialization time.

Bonus: I think this means we can get rid of the _legacy variants since this is also fixed for interpreter loading.

One quirk of this is it that you need to manually import torch from Python first otherwise there's an undefined symbol error likely due to not being dynamically linked with all of `torch` upfront. Shouldn't be an issue in opt mode.

Reviewed By: PaliC

Differential Revision: D37492006

fbshipit-source-id: 2e308c5d80ab38fd709bb3d012280bbb3641981f