Update base for Update on "add unit test to check for cuda"
Adds unit test to check if cuda is available when testing with gpu. Ends up it is pretty easy to build pytorch with USE_CUDA=1 but not actually use cuda or a nvidia gpu, this adds a siple test to check that we actually use cuda. This test gets run by executing the binary `build/test_deploy_gpu`.

Differential Revision: [D38430552](https://our.internmc.facebook.com/intern/diff/D38430552)

[ghstack-poisoned]