multipy/runtime: fix versioned symbol loading (ex torchaudio) (#140)
Summary:
This fixes issues where a dynamically loaded library (such as for torchaudio) has a dependency on a specific symbol version. TorchAudio requires `__log_finite` which is not a default symbol thus runtime fails to find it.

This updates the loader to parse the DT_VERNEED and DT_VERSYM entries in the ELF dynamic table to grab the required version string and pass it to `dlvsym`.

* torchaudio.so version info: https://gist.github.com/d4l3k/8724b0f3b1011a02287b825bed199486
* libm.so.6 `__log_finite` symbol: https://gist.github.com/d4l3k/3b4d9cd376338ec48109032dade2cdb7
* Symbol Versioning docs: https://refspecs.linuxfoundation.org/LSB_3.0.0/LSB-PDA/LSB-PDA.junk/symversion.html#VERNEEDFIG

Pull Request resolved: https://github.com/pytorch/multipy/pull/140

Test Plan: Enabled torchaudio in `test_compat.py`.

Reviewed By: priyaramani

Differential Revision: D38993735

Pulled By: d4l3k

fbshipit-source-id: 2aa29117ac34ba6a4e8722bd0e64473e9626f07d