Refactor py versioning in package tests yaml. (#185)
Summary:
We want to be able to pip install multipy runtime via setup.py. To do so, we need to install dependencies and use version numbers in the same way as runtime/nightly tests. This PR includes the py versioning changes to package tests' YAML.

Pull Request resolved: https://github.com/pytorch/multipy/pull/185

Test Plan: No changes to functionality with this PR.

Reviewed By: PaliC

Differential Revision: D39745163

Pulled By: anirbanr-fb-r2p

fbshipit-source-id: 47b7b7eae2566a1e837be46b475f97f3b5524369