Just Adding the InterpreterObj object and integrating it into interpreterImpl (#172)
Summary:
Pull Request resolved: https://github.com/pytorch/multipy/pull/172

This PR gets rid of the `objects_` variable in `ConcreteInterpreterSessionImpl`. Effectively,  this PR is a precursor to moving the logic of doing the py::bind function calls in `InterpreterSessionImpl` into `Obj` directly. We do this by creating a new struct called `InterpreterObj` who holds the underlying `py::object`. From here an `Obj` is simply a wrapper for the `InterpreterObj` as it was for `InterpreterSessionImpl` before.

Test Plan: Imported from OSS

Reviewed By: d4l3k

Differential Revision: D39583477

Pulled By: PaliC

fbshipit-source-id: dde512322cadbe04da660093eb45a26d569172ab