search recursively for static python (#290)
Summary:
Pull Request resolved: https://github.com/pytorch/multipy/pull/290

Somehow our [install tests](https://github.com/pytorch/multipy/actions/runs/3788974361/jobs/6444295267) and [runtime tests for python 3.7](https://github.com/pytorch/multipy/actions/runs/3788974394/jobs/6444296064) are broken on main due to what appears to be a linking issue.

The root cause of these breakages ended up being that the static python library was moved a bit deeper into the directory we thought it would be in when using pyenv. We now search recursively for it, so that if it is there we are able to use it. Furthermore, we also fail early is we cannot find static python in order to help prevent future breakages.

Test Plan: Imported from OSS

Reviewed By: d4l3k

Differential Revision: D42272568

Pulled By: PaliC

fbshipit-source-id: 6b164d9a562be1aefaf9446aa7175cfd8150dd66