Back out "Convert multipy embedded library exceptions to specific exception type" (#320)
Summary:
Pull Request resolved: https://github.com/pytorch/multipy/pull/320

X-link: https://github.com/pytorch/torchrec/pull/1153

Original commit changeset: 9cb3d1f17f54

Original Phabricator Diff: D44477860

In deploy, we do dlopen to link the .so each time when we create interpreter. So when interpreter manager is out of scope, we dlclose so we no longer have any symbols in the interpreter .so lib https://fburl.com/code/wlcp3r9o.

If we throw a custom exception class that defined in interpreter .so, then in the catch block, if interpreter manager goes out of scope, we don't recognize this exception class anymore and lead to a crash.

Reviewed By: zyan0

Differential Revision: D45629261

fbshipit-source-id: 5606f527e0952ee8a4184de25bd18b42faf772f3