Add caution message to readme before the repo gets archived. (#330)
Summary:
This is trying to clarify the maintenance status of multipy. In particular that it has been unmaintained for quite some time and there is no plans to invest in it going forward.
This is especially true with Free Threaded CPython becoming a reality and so the multithread limitations that multipy was trying to solve are becoming moot.
For users looking at LLM-like workloads, we also have much better solutions today (in particular vLLM) as they are more efficient and simpler than multipy.

You can see the rendering of the caution message at https://github.com/pytorch/multipy/tree/warn_dead

Pull Request resolved: https://github.com/pytorch/multipy/pull/330

Reviewed By: malfet

Differential Revision: D77989684

Pulled By: albanD

fbshipit-source-id: e24e1249da351a8415ec50819767372a50a5e065